queue update

+++ commands/music/play.js
@@ -1,6 +1,7 @@
 const { MessageEmbed } = require("discord.js");
 const ytsr = require('ytsr')
-const YTDL = require('ytdl-core');
+const ytdl = require('ytdl-core');
+const ytpl = require('ytpl');
 module.exports = {
     name: "play",
     category: "music",
@@ -18,14 +19,30 @@ module.exports = {
                 return;
             }
 
-            var filters = await ytsr.getFilters(mensaje);
-            filter = filters.get('Type').get('Video')
-
-            result = await ytsr(filter.url, {limit: 1}) 
-
-            let youtube = result.items[0]
-
-            Add(youtube.url, youtube.title, message.author.id, youtube.duration, youtube.bestThumbnail.url, youtube.author.name, youtube.author.url, message);
+            if(mensaje.includes("list=")){
+                ytpl(mensaje).then(async result => {
+                    const embed = new MessageEmbed()
+                        .setColor('#DD7F3F')
+                        .setTitle(`Added songs from the playlist: ${result.title}`)
+                        .setDescription(`${result.estimatedItemCount} songs added`)
+                        .setThumbnail(result.bestThumbnail.url)
+    
+                    await message.channel.send(embed);
+    
+                    result.items.forEach(song => {
+                        AddPlayList(song.url, song.title, message.author.id, song.duration, song.bestThumbnail.url, song.author.name, song.author.url, message)
+                    })
+                })
+            } else {
+                var filters = await ytsr.getFilters(mensaje);
+                filter = filters.get('Type').get('Video')
+    
+                result = await ytsr(filter.url, {limit: 1}) 
+    
+                let youtube = result.items[0]
+    
+                Add(youtube.url, youtube.title, message.author.id, youtube.duration, youtube.bestThumbnail.url, youtube.author.name, youtube.author.url, message);
+            }
             
         } 
         else 
@@ -47,16 +64,16 @@ module.exports = {
 
             const added = await message.channel.send(embed);
             
-            server.queue.push(url);
-            server.queueTitle.push(title);
-            server.queueThumbnail.push(thumbnail);
-            server.queueTime.push(timestamp);
-            server.queueRequestor.push(author);
-            server.queueAdded.push(added.id);
-            server.queueAuthorName.push(ytAuthor);
-            server.queueAuthorUrl.push(ytAuthorURL);
-
-            
+            server.music.push({
+                url: url,
+                title: title,
+                thumbnail: thumbnail,
+                timestamp: timestamp,
+                author: author,
+                added: added.id,
+                authorName:ytAuthor,
+                authorUrl: ytAuthorURL,
+            })
 
             if(!message.guild.me.voice.connection){
                 message.member.voice.channel.join()
@@ -97,8 +114,6 @@ async function Add(url, title, author, timestamp, thumbnail, ytAuthor, ytAuthorU
         authorUrl: ytAuthorURL,
     })
 
-    console.log(server.music)
-
     if(!message.guild.me.voice.connection){
         message.member.voice.channel.join()
         .then(connection=>{
@@ -111,7 +126,7 @@ async function Play(connection, message)
 {
     var server = servers[message.guild.id];
 
-    server.dispatcher = connection.play(YTDL(server.music[0].url, {highWaterMark: 1<<25, filter: "audioonly"}));
+    server.dispatcher = connection.play(ytdl(server.music[0].url, {highWaterMark: 1<<25, filter: "audioonly"}));
     
     var embed = new MessageEmbed()
             .setColor('#DD7F3F')
@@ -128,9 +143,11 @@ async function Play(connection, message)
         message.channel.messages.fetch(lastmsg).then(async msg => {
             msg.delete();
         });
-        message.channel.messages.fetch(server.music[0].added).then(async msg =>{
-            msg.delete();
-        })
+        if(server.music[0].added){
+            message.channel.messages.fetch(server.music[0].added).then(async msg =>{
+                msg.delete();
+            })
+        }
 
         server.music.shift()
 
@@ -146,4 +163,27 @@ async function Play(connection, message)
     });
 
     server.dispatcher.on('error', console.error);
+}
+
+async function AddPlayList(url, title, author, timestamp, thumbnail, ytAuthor, ytAuthorURL, message)
+{
+    var server = servers[message.guild.id];
+
+    server.music.push({
+        url: url,
+        title: title,
+        thumbnail: thumbnail,
+        timestamp: timestamp,
+        author: author,
+        added: null,
+        authorName:ytAuthor,
+        authorUrl: ytAuthorURL,
+    })
+
+    if(!message.guild.me.voice.connection){
+        message.member.voice.channel.join()
+        .then(connection=>{
+            Play(connection, message)
+        })
+    }
 }
